refactor: 💡 Changing directory structure and import destination
